# APISistemaVenta/SistemaVenta.API/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SistemaVenta.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Usuario/IniciarSesion": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/Guardar": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/Editar": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/Eliminar/{id}": {
      "delete": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Venta/Registrar": {
      "post": {
        "tags": [
          "Venta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VentaDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Venta/Historial": {
      "get": {
        "tags": [
          "Venta"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "buscarPor",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "numeroVenta",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "fechaInicio",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "fechaFin",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "vendedor",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "cliente",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Venta/Editar": {
      "put": {
        "tags": [
          "Venta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VentaDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Venta/Eliminar/{id}": {
      "delete": {
        "tags": [
          "Venta"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "DetalleVentaDTO": {
      "type": "object",
      "properties": {
        "idProducto": {
          "format": "int32",
          "type": "integer"
        },
        "descripcionProducto": {
          "type": "string"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "precioTexto": {
          "type": "string"
        },
        "totalTexto": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LoginDTO": {
      "type": "object",
      "properties": {
        "correo": {
          "type": "string"
        },
        "clave": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UsuarioDTO": {
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "nombreCompleto": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "idRol": {
          "format": "int32",
          "type": "integer"
        },
        "rolDescripcion": {
          "type": "string"
        },
        "clave": {
          "type": "string"
        },
        "esActivo": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "VentaDTO": {
      "type": "object",
      "properties": {
        "idVenta": {
          "format": "int32",
          "type": "integer"
        },
        "numeroDocumento": {
          "type": "string"
        },
        "tipoPago": {
          "type": "string"
        },
        "totalTexto": {
          "type": "string"
        },
        "esPagado": {
          "format": "int32",
          "type": "integer"
        },
        "cliente": {
          "type": "string"
        },
        "vendedor": {
          "type": "string"
        },
        "fechaRegistro": {
          "type": "string"
        },
        "detalleVenta": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DetalleVentaDTO"
          }
        }
      },
      "additionalProperties": false
    }
  }
}