# APISistemaVenta/SistemaVenta.DTO/LoginDTO.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace SistemaVenta.DTO
{
    public class LoginDTO
    {
        public string Correo { get; set; }
        public string Clave { get; set; }
    }
}
//Permite recibir la credencial del usuario  y autentificarla